created player controller component separated systems and renamed renamed playercontroller shared to prefabs shareddata

+++ Client/Assets/Scripts/Client/Components.cs
@@ -36,7 +36,7 @@ namespace Client
 
     public struct NetworkPlayerId : IComponentData
     {
-        public uint player;
+        public byte player;
         public NetworkConnection connection;
     }
